mobile improvements
