--- src/app/core/balance/balance.component.html
@@ -1,26 +1,0 @@
-<!-- TODO: add USDC withdraw -->
-<div
-  class="balance-withdraw"
-  [class.balance-withdraw--active]="usdcCanWithdraw"
-  (click)="clickWithdrawUsdc()"
-  [matTooltip]='
-    usdcCanWithdraw ?
-    "Click here to withdraw your USDC Balance" :
-    "You dont have any funds to withdraw yet"
-  '
->
-  <div class="balance-withdraw__label">
-    Withdrawable Balance
-  </div>
-  <div class="d-flex align-items-center">
-    <div class="balance-withdraw__icon">
-      <i class="fas fa-arrow-down"></i>
-    </div>
-    <div class="balance-withdraw__amount">
-      {{ usdcDisplayAvailable || 0 }}
-    </div>
-    <div class="balance-withdraw__currency">
-      USDC
-    </div>
-  </div>
-</div>
