--- src/app/core/balance/balance.component.scss
@@ -1,34 +1,0 @@
-@import '../../../scss/fonts';
-
-.balance-withdraw {
-  @include styled-font('Roboto', 700, 16px);
-  position: relative;
-  background: #8B8B8B36;
-  padding: 6px 10px 5px;
-  border: 1px solid var(--app-color-gray-50);
-  border-radius: 8px;
-  cursor: pointer;
-
-  &__icon,
-  &__amount {
-    color: var(--app-color-gray-50);
-  }
-  &__icon {
-    font-size: .95em;
-    margin-right: 6px;
-  }
-  &__amount {
-    margin-right: 5px;
-  }
-  &__label {
-    font-size: .7em;
-  }
-
-  &--active {
-    border: 1px solid var(--app-color-primary);
-    .balance-withdraw__icon,
-    .balance-withdraw__amount {
-      color: var(--app-color-primary);
-    }
-  }
-}
