+++ src/app/core/content-wrapper/content-wrapper.component.html
@@ -20,8 +20,7 @@
     </div>
     <div class="balance-container" *ngIf='hasAccount'>
       <div class="balance-row">
-        <app-component-balance [account]="account">
-        </app-component-balance>
+
       </div>
     </div>
     <div class="version-container">
