+++ src/app/core/header/icon-group-header/icon-group-header.component.html
@@ -54,7 +54,7 @@
 
 
   <app-notifications (notificationsCounter)="updateCounter($event)"></app-notifications>
-  <app-wallet-balances></app-wallet-balances>
+  <app-notification-wallet-balances></app-notification-wallet-balances>
   <app-wallet-withdraw [account]="account"></app-wallet-withdraw>
   <app-wallet-settings [account]="account"></app-wallet-settings>
 </div>
