+++ src/app/core/header/notification-wallet-balances/notification-wallet-balances.component.html
@@ -1,0 +1,6 @@
+<div
+  class="notifications-container"
+  [@anmNotifications]="viewDetail === 'balance' ? 'open' : 'closed'"
+>
+  <app-wallet-balances *ngIf="viewDetail === 'balance'"></app-wallet-balances>
+</div>
