+++ src/app/core/header/notification-wallet-balances/notification-wallet-balances.component.scss
@@ -1,0 +1,11 @@
+@import 'src/scss/variables';
+
+.notifications-container {
+  width: 182px;
+  max-height: 670px;
+  overflow: hidden;
+  position: absolute;
+  right: 208px;
+  @extend %roboto-regular;
+  @include transition();
+}
