--- src/app/core/header/wallet-balances/wallet-balances.component.scss
@@ -1,41 +1,0 @@
-@import './../../../../scss/variables';
-@import './../../../../scss/fonts';
-
-.notifications-container {
-  width: 182px;
-  max-height: 670px;
-  overflow: hidden;
-  position: absolute;
-  right: 208px;
-  @extend %roboto-regular;
-  @include transition();
-
-  ul {
-    li {
-      &:nth-child(1){
-        display: flex;
-        align-items: center;
-        height: 30px;
-        padding: 10px;
-        background-color: var(--app-color-gray-500);
-        border-bottom: 1px solid var(--app-color-gray-600);
-      }
-    }
-    .item {
-      padding: 11px 18px 11px 7px;
-      border-bottom: 1px solid var(--app-color-gray-600);
-      background-color: var(--app-color-gray-400);
-      color: white;
-      &__currency {
-        img {
-          width: 30px;
-          margin-right: 6px;
-        }
-        @include typography-overline(3);
-      }
-      &__amount {
-        @include typography-caption(1);
-      }
-    }
-  }
-}
