+++ src/app/dialogs/dialog-approve-contract/dialog-approve-contract.component.html
@@ -13,7 +13,7 @@
     *ngIf="startProgress"
     [start]="true"
     [end]="finishProgress"
-    (finish)="hideProgressbar()"
+    (finish)="finishApprove()"
   >
   </app-infinite-progress-bar>
 
@@ -32,114 +32,17 @@
         </div>
       </div>
 
-      <div
-        class="dialog-content__tabs row pt-md-2"
-        *ngIf="!onlyToken && !onlyAsset"
+      <app-approvals
+        [engine]="engine"
+        [onlyAddress]="onlyAddress"
+        [onlyToken]="onlyToken"
+        [onlyAsset]="onlyAsset"
+        (startProgress)="detectStartProgress()"
+        (endProgress)="detectEndProgress()"
+        (closeDialog)="detectCloseDialog()"
+        (setDialogDescription)="detectDialogDescription($event)"
       >
-        <div class="col-12 px-0">
-          <a
-            class="tab mr-md-3"
-            [class.tab--active]="engine === 'usdcEngine'"
-            (click)="clickEngine('usdcEngine')"
-          >
-            USDC Engine
-          </a>
-          <a
-            class="tab"
-            [class.tab--active]="engine === 'rcnEngine'"
-            (click)="clickEngine('rcnEngine')"
-          >
-            RCN Engine
-          </a>
-        </div>
-      </div>
-
-      <div class="dialog-content__content row mt-4 mt-md-0">
-        <div class="col-sm-12 px-md-0">
-
-          <ng-container *ngIf="tokens?.length && !onlyAsset">
-            <!-- <div class="title main-h2 mb-3 px-3 mx-1">
-              Tokens
-            </div> -->
-            <ng-container *ngFor="let token of tokens; let i = index">
-              <mat-accordion *ngIf="!onlyToken || onlyToken == token.address">
-                <mat-expansion-panel
-                  class="expansion-panel mat-elevation-z"
-                  [expanded]="
-                    onlyToken ||
-                    accordionStates.tokenApproves[token.address]
-                  "
-                  [hideToggle]="true"
-                  [disabled]="true"
-                >
-                  <mat-expansion-panel-header>
-                    <mat-panel-title>
-                      {{ token.name }}
-                    </mat-panel-title>
-                    <mat-slide-toggle
-                      (change)="clickToggleAccordion($event, 'tokenApproves', token.address)"
-                      [checked]="onlyToken || onlyAsset"
-                      color="primary"
-                    >
-                    </mat-slide-toggle>
-                  </mat-expansion-panel-header>
-                  <div
-                    class="checkbox"
-                    *ngFor="let operator of token.operators"
-                  >
-                    <mat-checkbox
-                      color="primary"
-                      [checked]="tokenApproves[token.address] && tokenApproves[token.address][operator.address]"
-                      (change)="clickCheck(operator, $event, token, 'token')"
-                      *ngIf="!onlyAddress || onlyAddress == operator.address"
-                      matTooltip="This will authorize the {{ operator.name }} contract to operate with {{ token.name }} on your behalf."
-                      matTooltipPosition="above"
-                    >
-                      Enable <b>{{ token.name }} {{ operator.action }}</b>.
-                    </mat-checkbox>
-                  </div>
-                </mat-expansion-panel>
-              </mat-accordion>
-            </ng-container>
-          </ng-container>
-
-          <ng-container *ngIf="assets?.length && !onlyToken">
-            <div class="title main-h2 mt-4 mb-3 px-3 mx-1">
-              ERC 721
-            </div>
-            <ng-container *ngFor="let asset of assets">
-              <mat-accordion *ngIf="!onlyAsset || onlyAsset == asset.address">
-                <mat-expansion-panel
-                  class="expansion-panel mat-elevation-z"
-                  [expanded]="onlyAsset"
-                >
-                  <mat-expansion-panel-header>
-                    <mat-panel-title>
-                      {{ asset.name }}
-                    </mat-panel-title>
-                  </mat-expansion-panel-header>
-                  <div
-                    class="checkbox"
-                    *ngFor="let operator of asset.operators"
-                  >
-                    <mat-checkbox
-                      color="primary"
-                      [checked]="assetApproves[asset.address] && assetApproves[asset.address][operator.address]"
-                      (change)="clickCheck(operator, $event, asset, 'asset')"
-                      *ngIf="!onlyAddress || onlyAddress == operator.address"
-                      matTooltip="This will authorize the {{ operator.name }} contract to operate with {{ asset.name }} on your behalf."
-                      matTooltipPosition="above"
-                    >
-                      Enable <b>{{ asset.name }} transactions</b>.
-                    </mat-checkbox>
-                  </div>
-                </mat-expansion-panel>
-              </mat-accordion>
-            </ng-container>
-          </ng-container>
-
-        </div>
-      </div>
+      </app-approvals>
     </div>
 
   </div>
