+++ src/app/shared/approvals/approvals.component.scss
@@ -1,0 +1,134 @@
+@import 'src/scss/fonts';
+
+.approvals {
+  &__tabs {
+    width: fit-content;
+    margin-left: auto;
+    margin-right: auto;
+    text-align: center;
+    border: 1px solid var(--app-color-gray-200);
+    border-radius: 18px;
+
+    .tab {
+      @include typography-body(2);
+      text-transform: uppercase;
+      display: inline-block;
+      cursor: pointer;
+      padding: 7px 20px;
+      letter-spacing: 1.4px;
+      &:not(.tab--active) {
+        color: #959595;
+      }
+      &:first-child {
+        border-right: 1px solid var(--app-color-gray-200);
+      }
+
+      /*
+      @include typography-body(2);
+      text-transform: uppercase;
+      cursor: pointer;
+      padding: 16px 20px;
+      letter-spacing: 1.4px;
+      &:not(.tab--active) {
+        color: #959595;
+        border-bottom: 1px solid transparent;
+      }
+      &--active {
+        border-bottom: 1px solid white;
+        cursor: default;
+      }
+      */
+    }
+  }
+
+  &__content {
+    .expansion-panel {
+      @include typography-headline(7);
+      background: black;
+      margin-bottom: 15px;
+      border-radius: 3px;
+      transition: .3s;
+      .mat-expansion-panel-header-title {
+        @include typography-body(2);
+        color: #747474;
+      }
+      &::ng-deep {
+        .mat-content {
+          overflow: visible;
+          align-items: center;
+        }
+        .mat-slide-toggle {
+          .mat-slide-toggle-bar {
+            height: 18px;
+            background: #4F526E;
+          }
+          .mat-slide-toggle-thumb-container {
+            top: -1px;
+            .mat-slide-toggle-thumb {
+              background: #9DA2CC;
+              box-shadow: 0px 3px 6px #00000084;
+            }
+          }
+          &.mat-checked {
+            .mat-slide-toggle-bar {
+              background: #242F8B;
+            }
+            .mat-slide-toggle-thumb-container {
+              .mat-slide-toggle-thumb {
+                background: #4155FF;
+              }
+            }
+          }
+        }
+      }
+      &.mat-expanded {
+        background: var(--app-color-gray-400);
+        .mat-expansion-panel-header-title {
+          color: white;
+        }
+      }
+    }
+
+    .checkbox {
+      margin-bottom: 12px;
+    }
+    &::ng-deep {
+      .mat-checkbox-label {
+        white-space: pre-wrap;
+      }
+      .mat-checkbox-inner-container {
+        width: 12px;
+        height: 12px;
+      }
+    }
+  }
+}
+
+@media (min-width: 992px) {
+  .approvals {
+    &__tabs {
+      width: 100%;
+      border: none;
+      border-radius: 0;
+      text-align: left;
+      margin-left: -15px;
+      margin-right: -15px;
+
+      .tab {
+        padding-top: 14px;
+        padding-bottom: 14px;
+        flex-shrink: 0;
+        &:first-child {
+          border-right: none;
+        }
+        &:not(.tab--active) {
+          border-bottom: 1px solid transparent;
+        }
+        &--active {
+          border-bottom: 1px solid white;
+          cursor: default;
+        }
+      }
+    }
+  }
+}
