+++ src/app/shared/wallet-avatar/wallet-avatar.component.scss
@@ -6,10 +6,6 @@
       opacity: .6;
     }
   }
-  img {
-    width: 25px;
-    height: 25px;
-  }
   &::after {
     content: '';
     position: absolute;
