+++ src/app/shared/wallet-avatar/wallet-avatar.component.spec.ts
@@ -1,7 +1,6 @@
 import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { MatSnackBarModule, MatDialogModule } from '@angular/material';
-import { SharedModule } from './../../../shared/shared.module';
 import { WalletAvatarComponent } from './wallet-avatar.component';
 
 describe('WalletAvatarComponent', () => {
@@ -10,7 +9,7 @@ describe('WalletAvatarComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      imports: [ MatSnackBarModule, MatDialogModule, SharedModule ],
+      imports: [ MatSnackBarModule, MatDialogModule ],
       declarations: [ WalletAvatarComponent ],
       schemas: [ CUSTOM_ELEMENTS_SCHEMA ]
     })
