+++ src/app/shared/wallet-balances/wallet-balances.component.html
@@ -1,0 +1,15 @@
+<ul>
+  <li>Wallet Balance</li>
+  <li
+    class="item d-flex justify-content-between align-items-center"
+    *ngFor="let balance of balances"
+  >
+    <div class="item__currency d-flex align-items-center">
+      <img [src]="balance.currency.img" [alt]="balance.currency.symbol">
+      {{ balance.currency.symbol }}
+    </div>
+    <div class="item__amount">
+      {{ balance.balance }}
+    </div>
+  </li>
+</ul>
