+++ src/app/views/my-account/balance/balance.component.html
@@ -1,0 +1,38 @@
+<!-- TODO: add USDC withdraw -->
+<div
+  class="balance"
+  [class.balance--active]="usdcCanWithdraw"
+  (click)="clickWithdrawUsdc()"
+  [matTooltip]='
+    usdcCanWithdraw ?
+    "Click here to withdraw your USDC Balance" :
+    "You dont have any funds to withdraw yet"
+  '
+>
+
+  <div class="balance__withdraw">
+    <div class="withdraw">
+      <div class="withdraw__label">
+        Withdrawable Balance
+      </div>
+      <div class="d-flex align-items-center">
+        <div class="withdraw__icon">
+          <i class="fas fa-arrow-down"></i>
+        </div>
+        <div class="withdraw__amount">
+          {{ usdcDisplayAvailable || 0 }}
+        </div>
+        <div class="withdraw__currency">
+          USDC
+        </div>
+      </div>
+    </div>
+  </div>
+
+  <div class="balance__cta" *ngIf="usdcCanWithdraw">
+    <div class="cta">
+      Click to withdraw
+    </div>
+  </div>
+
+</div>
