+++ src/app/views/my-account/balance/balance.component.scss
@@ -1,0 +1,47 @@
+@import 'src/scss/fonts';
+
+.balance {
+  position: relative;
+  width: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  @include styled-font('Roboto', 700, 16px);
+  background: #8B8B8B36;
+  padding: 6px 10px 5px;
+  border: 1px solid var(--app-color-gray-50);
+  border-radius: 8px;
+  cursor: pointer;
+
+  &__withdraw {
+    .withdraw {
+      &__icon,
+      &__amount {
+        color: var(--app-color-gray-50);
+      }
+      &__icon {
+        font-size: .95em;
+        margin-right: 6px;
+      }
+      &__amount {
+        margin-right: 5px;
+      }
+      &__label {
+        font-size: .7em;
+      }
+    }
+  }
+  &__cta {
+    .cta {
+      @include typography-subtitle(3);
+    }
+  }
+
+  &--active {
+    border: 1px solid var(--app-color-primary);
+    .withdraw__icon,
+    .withdraw__amount {
+      color: var(--app-color-primary);
+    }
+  }
+}
